// interactive-learning-v2/src/types/schema.ts
export type SectionType =
  | 'text'
  | 'tabs'
  | 'accordion'
  | 'timeline'
  | 'flashcards'
  | 'quiz'
  | 'fill-blank'
  | 'matching'
  | 'sorting'
  | 'checklist'
  | 'cloze';

export interface SectionBase {
  type: SectionType;
  title: string;
  _showNote?: boolean;
}
// ...
export interface FlashcardsSection extends SectionBase {
  type: 'flashcards';
  cards: Flashcard[];
}

export interface Flashcard {
  front: string;
  back: string;
}